Add UpiTurbo class for Turbo features
