+++ android/src/main/java/com/razorpay/razorpay_flutter/RazorpayDelegate.java
@@ -10,12 +10,8 @@ import com.google.gson.Gson;
 import com.razorpay.Checkout;
 import com.razorpay.CheckoutActivity;
 import com.razorpay.ExternalWalletListener;
-import com.razorpay.GenericPluginCallback;
 import com.razorpay.PaymentData;
 import com.razorpay.PaymentResultWithDataListener;
-import com.razorpay.UpiTurboLinkAccountResultListener;
-import com.razorpay.upi.Error;
-import com.razorpay.upi.UpiAccount;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -23,7 +19,6 @@ import org.json.JSONObject;
 import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
 
 import io.flutter.plugin.common.MethodChannel.Result;
@@ -50,10 +45,12 @@ public class RazorpayDelegate implements ActivityResultListener, ExternalWalletL
     private String packageName;
     private Checkout checkout;
     Gson gson ;
+    private UpiTurbo upiTurbo;
 
     public RazorpayDelegate(Activity activity) {
         this.activity = activity;
-        checkout = new Checkout().upiTurbo(activity);
+        upiTurbo = new UpiTurbo(activity);
+        /*checkout = new Checkout().upiTurbo(activity);*/
         this.gson = new Gson();
     }
 
@@ -189,87 +186,20 @@ public class RazorpayDelegate implements ActivityResultListener, ExternalWalletL
     }
 
     public void setKeyID(String keyId,  Result result){
-        checkout.setKeyID(keyId);
+        upiTurbo.setKeyID(keyId, result);
     }
 
     public void linkNewUpiAccount(String customerMobile, String color, Result result){
-        Log.d("RazorpayDelegate" , "customerMobile "+customerMobile +" color "+color);
-        this.pendingResult = result;
-        Map<String, Object> reply = new HashMap<>();
-        checkout.upiTurbo.linkNewUpiAccount(customerMobile, color , new GenericPluginCallback(){
-            @Override
-            public void onSuccess(@NonNull Object o) {
-                if (o instanceof List<?> && !((List<?>) o).isEmpty()) {
-                    reply.put("data", toJsonString(o));
-                } else {
-                    reply.put("data", "");
-                }
-                sendReply(reply);
-            }
-
-            @Override
-            public void onError(@NonNull JSONObject jsonObject) {
-                String errorCode = "AXIS_SDK_ERROR";
-                String errorDescription = "Something went wrong.Please try again.";
-                try {
-                    if (jsonObject.has("error")) {
-                        errorCode = jsonObject.getJSONObject("error").getString("code");
-                        errorDescription = jsonObject.getJSONObject("error").getString("description");
-                    }
-                } catch (Exception e) {
-                    Log.d("Exception", e.getMessage());
-                }
-                pendingResult.error(errorCode, errorDescription, jsonObject.toString());
-            }
-
-        });
+        upiTurbo.linkNewUpiAccount(customerMobile, color, result);
     }
 
 
     public void manageUpiAccounts(String customerMobile, String color, Result result){
-        this.pendingResult = result;
-        HashMap<Object, Object> reply = new HashMap<>();
-        checkout.upiTurbo.manageUpiAccounts(customerMobile, color , new GenericPluginCallback(){
-            @Override
-            public void onSuccess(@NonNull Object object) {
-            }
-
-            @Override
-            public void onError(@NonNull JSONObject jsonObject) {
-                String errorCode = "AXIS_SDK_ERROR";
-                String errorDescription = "Something went wrong.Please try again.";
-                try {
-                    if (jsonObject.has("error")) {
-                        errorCode = jsonObject.getJSONObject("error").getString("code");
-                        errorDescription = jsonObject.getJSONObject("error").getString("description");
-                    }
-                } catch (Exception e) {
-                    Log.d("Exception", e.getMessage());
-                }
-                pendingResult.error(errorCode, errorDescription, jsonObject.toString());
-            }
-        });
+        upiTurbo.manageUpiAccounts(customerMobile, color, result);
     }
 
     public  boolean isTurboPluginAvailable(Result result) {
-        this.pendingResult = result;
-
-        Map<String, Object> reply = new HashMap<>();
-        try {
-            Class.forName("com.razorpay.UpiTurboLinkAccountResultListener");
-            reply.put("isTurboPluginAvailable", true);
-            sendReply(reply);
-            return true;
-        } catch (ClassNotFoundException e) {
-            // Class not found, so it doesn't exist
-            reply.put("isTurboPluginAvailable", false);
-            sendReply(reply);
-            return false;
-        }
-    }
-
-    private String toJsonString(Object object){
-        return this.gson.toJson(object);
+        return upiTurbo.isTurboPluginAvailable(result);
     }
 
 }
