+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"fluttertoast","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/fluttertoast-8.2.2/","native_build":true,"dependencies":[]},{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":true,"dependencies":[]},{"name":"razorpay_flutter","path":"/Users/shibu.prasad/AndroidStudioProjects/razorpay-flutter/","native_build":true,"dependencies":["fluttertoast","package_info_plus"]}],"android":[{"name":"fluttertoast","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/fluttertoast-8.2.2/","native_build":true,"dependencies":[]},{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":true,"dependencies":[]},{"name":"razorpay_flutter","path":"/Users/shibu.prasad/AndroidStudioProjects/razorpay-flutter/","native_build":true,"dependencies":["fluttertoast","package_info_plus"]}],"macos":[{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":true,"dependencies":[]}],"linux":[{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":false,"dependencies":[]}],"windows":[{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":false,"dependencies":[]}],"web":[{"name":"fluttertoast","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/fluttertoast-8.2.2/","dependencies":[]},{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","dependencies":[]}]},"dependencyGraph":[{"name":"fluttertoast","dependencies":[]},{"name":"package_info_plus","dependencies":[]},{"name":"razorpay_flutter","dependencies":["fluttertoast","package_info_plus"]}],"date_created":"2023-11-24 11:54:59.199320","version":"3.13.9"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"fluttertoast","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/fluttertoast-8.2.2/","native_build":true,"dependencies":[]},{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":true,"dependencies":[]},{"name":"razorpay_flutter","path":"/Users/shibu.prasad/AndroidStudioProjects/razorpay-flutter/","native_build":true,"dependencies":["fluttertoast","package_info_plus"]}],"android":[{"name":"fluttertoast","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/fluttertoast-8.2.2/","native_build":true,"dependencies":[]},{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":true,"dependencies":[]},{"name":"razorpay_flutter","path":"/Users/shibu.prasad/AndroidStudioProjects/razorpay-flutter/","native_build":true,"dependencies":["fluttertoast","package_info_plus"]}],"macos":[{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":true,"dependencies":[]}],"linux":[{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":false,"dependencies":[]}],"windows":[{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","native_build":false,"dependencies":[]}],"web":[{"name":"fluttertoast","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/fluttertoast-8.2.2/","dependencies":[]},{"name":"package_info_plus","path":"/Users/shibu.prasad/.pub-cache/hosted/pub.dev/package_info_plus-4.0.1/","dependencies":[]}]},"dependencyGraph":[{"name":"fluttertoast","dependencies":[]},{"name":"package_info_plus","dependencies":[]},{"name":"razorpay_flutter","dependencies":["fluttertoast","package_info_plus"]}],"date_created":"2023-12-06 13:19:25.257691","version":"3.13.9"}
