+++ example/android/app/build.gradle
@@ -93,6 +93,6 @@ dependencies {
     implementation 'com.google.android.gms:play-services-auth:20.2.0'
     implementation 'com.google.android.gms:play-services-auth-api-phone:18.0.1'
 
-    implementation fileTree(include: ['*.aar'], dir: '/Users/shibu.prasad/Desktop/FlutterStandardWithTPV/')
+    //implementation fileTree(include: ['*.aar'], dir: '/Users/shibu.prasad/Desktop/FlutterStandardWithTPV/')
     //implementation fileTree(include: ['*.aar'], dir: '/Users/atif.qamar/Desktop/StandardCheckout/CheckoutLib/')
 }
